+++ changelog.md
@@ -3,6 +3,13 @@
 Versioning is loose and lax, in fact there are no actual versions.  
 But here is the list of changes made over time, sorted by 'release' date.
 
+## March 29, 2026
+- New: Separate cleanup function (cache_delete()) for cache
+- Change: Last build date for feeds now matches newest/latest item in the feed
+- Change: Improved error handling when fetching feeds
+- Fix: Youtube video publish date now processed as string, not an int
+- Fix: removed unused feed.php
+
 ## March 23, 2026
 - Change: Fetched variables are now cast as a string/int/bool
 - Change: Merged YT and EZTV field on subscribe.php for easier use